% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/say_hello.R
\name{say_hello}
\alias{say_hello}
\title{Say Hello}
\usage{
say_hello(name)
}
\arguments{
\item{name}{A character string containing the name of the person to greet.}
}
\value{
A greeting string.
}
\description{
A simple function that returns a personalized greeting.
}
\examples{
say_hello("Jellia")
}
